+++ private/pkg/protosource/protosource.go
@@ -295,6 +295,12 @@ type MessageRange interface {
 	Message() Message
 }
 
+// ExtensionRange represents an extension range in Messages.
+type ExtensionRange interface {
+	MessageRange
+	OptionExtensionDescriptor
+}
+
 // Enum is an enum descriptor.
 type Enum interface {
 	NamedDescriptor
@@ -336,6 +342,7 @@ type Message interface {
 	Fields() []Field
 	Extensions() []Field
 	Oneofs() []Oneof
+	ExtensionRanges() []ExtensionRange
 	ExtensionMessageRanges() []MessageRange
 	ReservedMessageRanges() []MessageRange